% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvelo.R
\name{run_scvelo_example}
\alias{run_scvelo_example}
\title{Run scvelo() example}
\usage{
run_scvelo_example()
}
\value{
The output of `scvelo()`.
}
\description{
This function is a hotfix for an issue running example code during
`R CMD check`.
}
\examples{
run_scvelo_example()
}
