% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\name{scveloCondaInfo}
\alias{scveloCondaInfo}
\title{List Conda Packages Information}
\usage{
scveloCondaInfo()
}
\value{
The data.frame returned by \code{basilisk::listPackages()}.
}
\description{
List package names and versions in the \pkg{scVelo} Conda environment.
}
\examples{
scveloCondaInfo()
}
