% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velociraptor-package.R
\docType{package}
\name{velociraptor-package}
\alias{velociraptor}
\alias{velociraptor-package}
\title{velociraptor: Toolkit for Single-Cell Velocity}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package provides Bioconductor-friendly wrappers for RNA velocity calculations in single-cell RNA-seq data.
    We use the basilisk package to manage Conda environments,
    and the zellkonverter package to convert data structures between SingleCellExperiment (R) and AnnData (Python).
    The information produced by the velocity methods is stored in the various components of the SingleCellExperiment class.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinrue/velociraptor}
  \item Report bugs at \url{https://github.com/kevinrue/velociraptor/issues}
}

}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevinrue67@gmail.com} (\href{https://orcid.org/0000-0003-3899-3872}{ORCID})

Authors:
\itemize{
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID})
  \item Charlotte Soneson \email{charlottesoneson@gmail.com} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})
}

}
\keyword{internal}
