% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocyto.R
\name{velocyto}
\alias{velocyto}
\alias{velocyto,ANY-method}
\alias{velocyto,SummarizedExperiment-method}
\title{RNA velocity with \pkg{velocyto}}
\usage{
velocyto(x, ...)

\S4method{velocyto}{ANY}(
  x,
  subset.row = NULL,
  sf.spliced = NULL,
  sf.unspliced = NULL,
  use.theirs = FALSE,
  velocyto.params = list()
)

\S4method{velocyto}{SummarizedExperiment}(
  x,
  ...,
  assay.spliced = "spliced",
  assay.unspliced = "unspliced",
  assay.ambiguous = NULL
)
}
\arguments{
\item{x}{A list of count matrices of the same dimensions, with genes in rows and cells in samples.
This should contain \code{"spliced"} and \code{"unspliced"} (and possibly \code{"ambiguous"}, see Details).

Alternatively, a \linkS4class{SummarizedExperiment} object containing three such matrices among its assays.}

\item{...}{For the generic, further arguments to pass to specific methods.
For the SummarizedExperiment and SingleCellExperiment methods, further arguments to pass to the ANY method.}

\item{subset.row}{A character, integer or logical vector specifying the genes to use for the velocity calculations.
Defaults to all genes.}

\item{sf.spliced}{A numeric vector containing size factors for the spliced counts for each cell.
Defaults to \code{\link{librarySizeFactors}} on the \code{"spliced"} matrix in \code{x}.}

\item{sf.unspliced}{A numeric vector containing size factors for the unspliced counts for each cell.
Defaults to \code{\link{librarySizeFactors}} on the \code{"unspliced"} matrix in \code{x}.}

\item{use.theirs}{Logical scalar indicating whether \pkg{velocyto}'s gene filtering and normalization should be used.}

\item{velocyto.params}{List of lists containing arguments for individual \pkg{velocyto} functions, see details below.}

\item{assay.spliced}{An integer scalar or string specifying the assay of \code{x} containing the spliced counts.}

\item{assay.unspliced}{An integer scalar or string specifying the assay of \code{x} containing the unspliced counts.}

\item{assay.ambiguous}{An integer scalar or string specifying the assay \code{x} containing the ambiguous counts.
If \code{NULL}, no attempt is made to extract this count matrix, see Details.}
}
\value{
A \linkS4class{SummarizedExperiment} is returned where each column corresponds to a cell in \code{x}.
Each row corresponds to a feature after subsetting if \code{subset.row} is provided;
or after feature selection within \pkg{velocyto} if \code{use.theirs=TRUE}.
This object contains the following assays:
\itemize{
\item \code{velocity}, a matrix containing per-cell velocity vectors in each dimension.
\item \code{extrapolated}, a matrix containing the predicted location of each cell along its velocity vector.
}
}
\description{
Perform RNA velocity calculations using the original implementation
of the steady-state model by La Manno et al. (2018).
}
\details{
This function uses the \pkg{velocyto} Python package (\url{https://pypi.org/project/velocyto/}) for RNA velocity calculations.
All matrices are saved to a temporary \code{.loom} file that is subsequently read by \pkg{velocyto} functions.
This is necessary as \pkg{velocyto} itself does not provide an entry point for matrices that are already in memory.

For consistency with other Bioconductor workflows, we perform as many standard steps in R as we can
before starting the velocity calculations with \pkg{scVelo}.
This involves:
\enumerate{
\item Size factor calculations, using \code{\link{librarySizeFactors}} on entries of \code{x}.
If \code{sf.*} arguments are supplied, these are used directly.
\item Subsetting all matrices to \code{subset.row}, most typically to a subset of interest, e.g., highly variable genes.
Note that, if set, any subsetting is done \emph{after} size factor calculations so that library sizes are correctly computed.
In addition, the \code{.loom} file will only contain the specified subset to reduce the I/O overhead.
}

Users can set \code{use.theirs=TRUE} to directly use the entire \pkg{velocyto} normalization and filtering pipeline.
This ignores all of the size factor arguments (\code{sf.*}) as well as \code{subset.row}.

All individual \pkg{velocyto} function calls are performed with their default values,
but this can be changed by including named parameter lists in \code{velocyto.params}.
Supported arguments are:
\itemize{
\item \code{score_cv_vs_mean}, for CV-mean calculations if \code{use.theirs=TRUE}.
\item \code{filter_genes}, for feature selection if \code{use.theirs=TRUE}.
\item \code{normalize}, for normalization if \code{use.theirs=TRUE}.
\item \code{perform_PCA}
\item \code{knn_imputation}
\item \code{fit_gammas}
\item \code{predict_U}
\item \code{calculate_velocity}
\item \code{calculate_shift}
\item \code{extrapolate_cell_at_t}
}

It does not seem that the \code{"ambiguous"} entry in \code{x} is used for anything,
but it is nonetheless necessary to supply \emph{some} value as \pkg{velocyto} will fail otherwise.
If such an entry is not present in \code{x}, an empty sparse matrix will be created instead.

Upon first use, this function will instantiate a conda environment containing the \pkg{velocyto} package.
This is done via the \pkg{basilisk} package - see the documentation for that package for trouble-shooting.
}
\examples{
# Using mock data to demonstrate the process:
library(scuttle)
sce1 <- mockSCE()
sce2 <- mockSCE()

spliced <- counts(sce1)
unspliced <- counts(sce2)

out <- velocyto(list(spliced=spliced, unspliced=unspliced))

}
\references{
La Manno G, Soldatov R, Zeisel A et al. (2018).
RNA velocity of single cells.
\emph{Nature} 560, 494-498
}
\seealso{
\code{\link{scvelo}} with \code{mode="steady-state"},
which provides an alternative implementation of steady-state model.
}
