% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvelo.R
\name{scvelo}
\alias{scvelo}
\alias{scvelo,ANY-method}
\alias{scvelo,SummarizedExperiment-method}
\title{RNA velocity calculations with \pkg{scvelo}}
\usage{
scvelo(x, ...)

\S4method{scvelo}{ANY}(x)

\S4method{scvelo}{SummarizedExperiment}(x, ..., assay.spliced = "spliced", assay.unspliced = "unspliced")
}
\arguments{
\item{x}{A list of two matrices of the same dimensions where genes are in rows and cells are in columns.
The first matrix should contain spliced counts and te second matrix should contain unspliced counts.

Alternatively, a \linkS4class{SummarizedExperiment} object containing two such matrices in its assays.}

\item{...}{For the generic, further arguments to pass to specific methods.
For the SummarizedExperiment method, further arguments to pass to the ANY method.}

\item{assay.spliced}{An integer scalar or string specifying the assay of \code{x} containing the spliced counts.}

\item{assay.unspliced}{An integer scalar or string specifying the assay of \code{x} containing the unspliced counts.}
}
\value{
By default, a \linkS4class{DataFrame} is returned containing:
}
\description{
Perform RNA velocity calculations with the \pkg{scvelo} package.
}
\details{
This function uses the \pkg{scvelo} package (\url{https://pypi.org/project/scvelo/}) to perform RNA velocity calculations.
The main difference from \code{\link{velocyto}} is that it does not rely on the presence of observed steady state populations.

Upon first use, this function will instantiate a conda environment containing the \pkg{scvelo} package.
}
\examples{
# Using mock data to demonstrate the process:
library(scuttle)
sce1 <- mockSCE()
sce2 <- mockSCE()

spliced <- counts(sce1)
unspliced <- counts(sce2)

out <- scvelo(list(spliced, unspliced))

}
\author{
Aaron Lun
}
