% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvelo.R
\name{scvelo}
\alias{scvelo}
\alias{scvelo,ANY-method}
\alias{scvelo,SummarizedExperiment-method}
\alias{scvelo,SingleCellExperiment-method}
\title{RNA velocity calculations with \pkg{scVelo}}
\usage{
scvelo(x, ...)

\S4method{scvelo}{ANY}(
  x,
  subset.row = NULL,
  norm = FALSE,
  sf.x = NULL,
  sf.spliced = NULL,
  sf.unspliced = NULL,
  use.theirs = FALSE,
  mode = c("steady_state", "deterministic", "stochastic", "dynamical"),
  scvelo.params = list(),
  dimred = NULL,
  ncomponents = 30,
  BPPARAM = SerialParam(),
  BSPARAM = bsparam()
)

\S4method{scvelo}{SummarizedExperiment}(
  x,
  ...,
  assay.x = "counts",
  assay.spliced = "spliced",
  assay.unspliced = "unspliced",
  sf.x = sizeFactors(x),
  sf.spliced = NULL,
  sf.unspliced = NULL
)

\S4method{scvelo}{SingleCellExperiment}(
  x,
  ...,
  assay.x = "counts",
  assay.spliced = "spliced",
  assay.unspliced = "unspliced",
  sf.x = sizeFactors(x),
  sf.spliced = NULL,
  sf.unspliced = NULL,
  dimred = "PCA"
)
}
\arguments{
\item{x}{A list of three matrices of the same dimensions where genes are in rows and cells are in columns.
The first matrix should contain the counts used for calculating reduced dimension representations and cell-to-cell distances.
The second matrix should contain spliced counts and the third matrix should contain unspliced counts.
If \code{norm=TRUE}, all matrices should contain normalized (but not log-transformed) expression values instead.

Alternatively, a \linkS4class{SummarizedExperiment} object containing three such matrices in its assays.}

\item{...}{For the generic, further arguments to pass to specific methods.
For the SummarizedExperiment method, further arguments to pass to the ANY method.}

\item{subset.row}{A character, integer or logical vector specifying the genes to use for the velocity calculations.
Defaults to all genes but is most typically set to a subset of interesting genes, e.g., highly variable genes.
Note that, if set, any subsetting is done \emph{after} normalization so that library sizes are correctly computed.}

\item{norm}{Logical scalar indicating whether the matrices in \code{x} are already normalized.
In such cases, \code{sf.x}, \code{sf.spliced} and \code{sf.unspliced} are ignored.}

\item{sf.x}{A numeric vector containing size factors for the counts in \code{x} for each cell.
Defaults to \code{\link{librarySizeFactors}} on the \code{x} count matrix.}

\item{sf.spliced}{A numeric vector containing size factors for the spliced counts for each cell.
Defaults to \code{\link{librarySizeFactors}} on the spliced matrix.}

\item{sf.unspliced}{A numeric vector containing size factors for the unspliced counts for each cell.
Defaults to \code{\link{librarySizeFactors}} on the unspliced matrix.}

\item{use.theirs}{Logical scalar indicating whether \pkg{scVelo}'s gene filtering and normalization should be used,
in which case \code{sf.counts}, \code{sf.spliced}, \code{sf.unspliced} and \code{subset.row} are ignored.}

\item{mode}{String specifying the method to use to estimate the transcriptional dynamics.}

\item{scvelo.params}{List of lists, providing arguments for scVelo functions.}

\item{dimred}{A low-dimensional representation of the cells (N x k matrix, where N is the number of cells) that will be used to find the nearest neighbors for the moment calculations.
Alternatively, if \code{x} is a \linkS4class{SingleCellExperiment}, a character scalar with the name of a reduced dimension representation in \code{x}.}

\item{ncomponents}{Numeric scalar indicating the number of principal components to obtain, if \code{use.theirs} is FALSE and no \code{dimred} is supplied.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether the PCA calculations should be parallelized.}

\item{BSPARAM}{A \linkS4class{BiocSingularParam} object specifying which algorithm should be used to perform the PCA.}

\item{assay.x}{An integer scalar or string specifying the assay of \code{x} containing the counts that will be used for dimension reduction and distance calculations.}

\item{assay.spliced}{An integer scalar or string specifying the assay of \code{x} containing the spliced counts.}

\item{assay.unspliced}{An integer scalar or string specifying the assay of \code{x} containing the unspliced counts.}
}
\value{
A \linkS4class{SummarizedExperiment} is returned containing the output of the velocity calculations.
TODO: SOME DOCUMENTATION REQUIRED.
}
\description{
Perform RNA velocity calculations with the \pkg{scVelo} package.
}
\details{
This function uses the \pkg{scVelo} package (\url{https://pypi.org/project/scvelo/}) to perform RNA velocity calculations.
The main difference from \pkg{velocyto} is that the dynamical model of \pkg{scVelo} does not rely on the presence of observed steady state populations,
which may improve the reliability of the velocity calculations in general applications.

For consistency with other Bioconductor packages, we perform scaling normalization and subsetting
before starting the velocity calculations with \pkg{scVelo}.
This allows us to guarantee that, e.g., the spliced log-expression matrix of HVGs is the same as that used in other
applications like clustering or trajectory reconstruction.
Nonetheless, one can set \code{use.theirs=TRUE} to directly use the entire \pkg{scVelo} normalization and filtering pipeline.

Upon first use, this function will instantiate a conda environment containing the \pkg{scVelo} package.

Additional arguments to \pkg{scVelo} functions are provided via the \code{scvelo.params} list.
The arguments for a given function are given as a named list.
The following function names are currently recognized:
\itemize{
\item filter_and_normalize
\item moments
\item recover_dynamics
\item velocity
\item velocity_graph
\item velocity_pseudotime
\item latent_time
\item velocity_confidence
}
For more information about the arguments of these functions, see the \pkg{scVelo} documentation.
}
\examples{
# Using mock data to demonstrate the process:
library(scuttle)
sce1 <- mockSCE()
sce2 <- mockSCE()

spliced <- counts(sce1)
unspliced <- counts(sce2)

out <- scvelo(list(spliced, spliced, unspliced), use.theirs=TRUE)

}
\author{
Aaron Lun, Charlotte Soneson
}
