% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velociraptor-package.R
\docType{package}
\name{velociraptor-package}
\alias{velociraptor}
\alias{velociraptor-package}
\title{velociraptor: Toolkit for Single-Cell Velocity}
\description{
Functionality for processing single-cell velocity data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinrue/velociraptor}
  \item Report bugs at \url{https://github.com/kevinrue/velociraptor/issues}
}

}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevin.rue-albrecht@ndlcs.ox.ac.uk} (\href{https://orcid.org/0000-0003-3899-3872}{ORCID})

Authors:
\itemize{
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID})
}

}
\keyword{internal}
